# MACRO: expand `&item*` pattern