--- a/src/main/kotlin/org/rust/lang/core/macros/LazyStatic.kt
+++ b/src/main/kotlin/org/rust/lang/core/macros/LazyStatic.kt
@@ -7,16 +7,19 @@ package org.rust.lang.core.macros
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.tree.IElementType
-import org.rust.lang.core.psi.*
+import org.rust.lang.core.psi.RsConstant
+import org.rust.lang.core.psi.RsElementTypes
+import org.rust.lang.core.psi.RsMacroCall
+import org.rust.lang.core.psi.RsTt
 import org.rust.lang.core.psi.ext.elementType
 
 fun expandLazyStatic(call: RsMacroCall): ExpansionResult? {
     val arg = call.macroArgument?.tt ?: return null
     val lazyStaticCall = parseLazyStaticCall(arg) ?: return null
-    return RsCodeFragmentFactory(call.project)
-        .createExpandedItem<RsConstant>(
-            "${if (lazyStaticCall.pub) "pub " else ""}static ${lazyStaticCall.identifier}: ${lazyStaticCall.type} = &${lazyStaticCall.expr};"
-        )
+    return createInMemoryPsi<RsConstant>(
+        call.project,
+        "${if (lazyStaticCall.pub) "pub " else ""}static ${lazyStaticCall.identifier}: ${lazyStaticCall.type} = &${lazyStaticCall.expr};"
+    )
 }
 
 private data class LazyStaticCall(