--- a/src/main/kotlin/org/rust/lang/core/psi/ext/RsExternCrateItem.kt
+++ b/src/main/kotlin/org/rust/lang/core/psi/ext/RsExternCrateItem.kt
@@ -9,6 +9,7 @@ import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 import com.intellij.psi.stubs.IStubElementType
 import org.rust.ide.icons.RsIcons
+import org.rust.lang.core.macros.ExpansionResult
 import org.rust.lang.core.psi.RsExternCrateItem
 import org.rust.lang.core.psi.RsPsiImplUtil
 import org.rust.lang.core.resolve.ref.RsExternCrateReferenceImpl
@@ -31,6 +32,8 @@ abstract class RsExternCrateItemImplMixin : RsStubbedNamedElementImpl<RsExternCr
     override val isPublic: Boolean get() = RsPsiImplUtil.isPublic(this, stub)
 
     override fun getIcon(flags: Int) = RsIcons.CRATE
+
+    override fun getContext() = ExpansionResult.getContextImpl(this)
 }
 
 val RsExternCrateItem.hasMacroUse: Boolean get() =