--- a/src/main/kotlin/org/rust/lang/core/psi/ext/RsForeignModItem.kt
+++ b/src/main/kotlin/org/rust/lang/core/psi/ext/RsForeignModItem.kt
@@ -8,6 +8,7 @@ package org.rust.lang.core.psi.ext
 import com.intellij.lang.ASTNode
 import com.intellij.psi.stubs.IStubElementType
 import com.intellij.psi.util.PsiTreeUtil
+import org.rust.lang.core.macros.ExpansionResult
 import org.rust.lang.core.psi.RsForeignModItem
 import org.rust.lang.core.psi.RsOuterAttr
 import org.rust.lang.core.stubs.RsPlaceholderStub
@@ -23,4 +24,6 @@ abstract class RsForeignModItemImplMixin : RsStubbedElementImpl<RsPlaceholderStu
         get() = PsiTreeUtil.getChildrenOfTypeAsList(this, RsOuterAttr::class.java)
 
     override val isPublic: Boolean get() = false // visibility does not affect foreign mods
+
+    override fun getContext() = ExpansionResult.getContextImpl(this)
 }