--- a/src/main/kotlin/org/rust/lang/core/psi/ext/RsFunction.kt
+++ b/src/main/kotlin/org/rust/lang/core/psi/ext/RsFunction.kt
@@ -10,6 +10,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.stubs.IStubElementType
 import org.rust.ide.icons.RsIcons
 import org.rust.ide.icons.addTestMark
+import org.rust.lang.core.macros.ExpansionResult
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.stubs.RsFunctionStub
 import org.rust.lang.core.types.ty.Ty
@@ -133,4 +134,6 @@ abstract class RsFunctionImplMixin : RsStubbedNamedElementImpl<RsFunctionStub>, 
             else -> RsIcons.METHOD
         }
     }
+
+    override fun getContext(): RsElement = ExpansionResult.getContextImpl(this)
 }