… src/main/kotlin/org/rust/lang/core/psi/ext/RsMacroCall.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsModDeclItem.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsModItem.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsStructItem.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsTraitItem.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsTypeAlias.kt
… src/main/kotlin/org/rust/lang/core/psi/ext/RsUseItem.kt
… src/main/kotlin/org/rust/lang/core/stubs/StubImplementations.kt
… src/main/kotlin/org/rust/stdext/Collections.kt
… src/test/kotlin/org/rust/lang/core/resolve/RsMacroExpansionResolveTest.kt