if_std in futures-rs and failure now works