@matklad 2017-11-30
r? @vlad20012 

Have not measure the performance impact, will do now. 

I think the implementation is more or less solid, but I absolutely hate this `ItemCache` stuff now :(