@matklad 2017-11-30
Performance regressed for ~ 30% 😞 

<details>

```
BEFORE:

highlighting mysql_async:
collecting  : 4 ms
resolve     : 2817 ms
highlighting: 184 ms
3005 ms total.

highlighting Cargo:
collecting  : 4 ms
resolve     : 850 ms
highlighting: 91 ms
945 ms total.

AFTER:

highlighting mysql_async:
collecting  : 3 ms
resolve     : 4139 ms
highlighting: 198 ms
4340 ms total.

highlighting Cargo:
collecting  : 4 ms
resolve     : 1114 ms
highlighting: 98 ms
1216 ms total.
```

</details>