@matklad 2017-12-01
Added a cache (which helped a bit!) and make sure we use a single instance of RsPsiFactory (not that it should impact performance, but it should nudge us into grouping macro-related stuff together). 

So, this is ready for review I guess :)