@Undin 2017-12-01
If open `future-rs` project and go to `src/stream/collect.rs` all items from other files can't be resolved. But if you comment usage of `if_std` macros in `src/stream/mod.rs` resolve works