@matklad 2017-12-01
Hm, yeah, that one is going to be tough =/ The problem is that we can't determine parent module in this case, because `mod collect;` mod decl is inside `if_std!`, so it is not present in indexes ...